+++ src/features/group/tools/index.ts
@@ -6,6 +6,7 @@ import { pimGroupApprovalListTool } from "./pim-group-approval-list.js";
 import { pimGroupApprovalReviewTool } from "./pim-group-approval-review.js";
 import { pimGroupDeactivateTool } from "./pim-group-deactivate.js";
 import { pimGroupEligibleListTool } from "./pim-group-eligible-list.js";
+import { pimGroupRequestCancelTool } from "./pim-group-request-cancel.js";
 import { pimGroupRequestListTool } from "./pim-group-request-list.js";
 import { pimGroupRequestTool } from "./pim-group-request.js";
 
@@ -15,6 +16,7 @@ export {
   pimGroupApprovalReviewTool,
   pimGroupDeactivateTool,
   pimGroupEligibleListTool,
+  pimGroupRequestCancelTool,
   pimGroupRequestListTool,
   pimGroupRequestTool,
 };
@@ -24,6 +26,7 @@ export const GROUP_TOOLS: readonly AnyTool[] = [
   pimGroupActiveListTool,
   pimGroupRequestListTool,
   pimGroupRequestTool,
+  pimGroupRequestCancelTool,
   pimGroupDeactivateTool,
   pimGroupApprovalListTool,
   pimGroupApprovalReviewTool,